pyproject.toml
[tool.poetry]
name = "restaurant-data"
version = "0.1.0"
description = ""
authors = ["Nicolas Jonckheere <nicolasjonckheere22@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
google-cloud-storage = "^2.15.0"
google-cloud-bigquery = "^3.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
